+++ curves/benches/field_arith.rs
@@ -7,7 +7,9 @@
 //!
 //!     cargo bench --bench field_arith
 
-use criterion::{black_box, criterion_group, criterion_main, Criterion, Throughput};
+use std::hint::black_box;
+
+use criterion::{criterion_group, criterion_main, Criterion, Throughput};
 use ff::Field;
 use midnight_curves::*;
 use rand_core::{RngCore, SeedableRng};
@@ -83,9 +85,8 @@ fn bench_field_arithmetic<F: Field>(c: &mut Criterion, name: &'static str) {
         bencher.iter(|| black_box(&a).sub(black_box(&b)))
     });
 
-    ret = a;
-    group.bench_function(format!("{}_sub", name), |bencher| {
-        bencher.iter(|| black_box(&mut ret).sub(black_box(&b)))
+    group.bench_function(format!("{}_sub_assign", name), |bencher| {
+        bencher.iter(|| black_box(&mut ret).sub_assign(black_box(&b)))
     });
 
     group.bench_function(format!("{}_double", name), |bencher| {
@@ -100,9 +101,8 @@ fn bench_field_arithmetic<F: Field>(c: &mut Criterion, name: &'static str) {
         bencher.iter(|| black_box(&a).mul(black_box(&b)))
     });
 
-    ret = a;
-    group.bench_function(format!("{}_mul", name), |bencher| {
-        bencher.iter(|| black_box(&mut ret).mul(black_box(&b)))
+    group.bench_function(format!("{}_mul_assign", name), |bencher| {
+        bencher.iter(|| black_box(&mut ret).mul_assign(black_box(&b)))
     });
 
     group.bench_function(format!("{}_square", name), |bencher| {