+++ curves/src/g1.rs
@@ -344,8 +344,7 @@ impl G1Affine {
 
     /// Attempts to deserialize an uncompressed element.
     fn from_uncompressed(bytes: &[u8; UNCOMPRESSED_SIZE]) -> CtOption<Self> {
-        G1Affine::from_uncompressed_unchecked(bytes)
-            .and_then(|p| CtOption::new(p, p.is_on_curve() & p.is_torsion_free()))
+        G1Affine::from_uncompressed_unchecked(bytes).and_then(|p| CtOption::new(p, p.is_on_curve()))
     }
 
     /// Attempts to deserialize an uncompressed element, not checking if the
@@ -711,6 +710,19 @@ impl PrimeGroup for G1Projective {}
 
 impl Curve for G1Projective {
     type AffineRepr = G1Affine;
+
+    /// Converts a batch of projective elements into affine elements. This
+    /// function will panic if `p.len() != q.len()`.
+    fn batch_normalize(p: &[Self], q: &mut [Self::AffineRepr]) {
+        assert_eq!(p.len(), q.len());
+        let points = unsafe { std::slice::from_raw_parts(p.as_ptr() as *const blst_p1, p.len()) };
+
+        p1_affines::from(points)
+            .as_slice()
+            .iter()
+            .zip(q.iter_mut())
+            .for_each(|(val, res)| *res = G1Affine(*val));
+    }
 
     fn to_affine(&self) -> Self::AffineRepr {
         self.into()