+++ proofs/src/plonk/verifier.rs
@@ -1,4 +1,4 @@
-use std::iter;
+use std::{hash::Hash, iter};
 
 use ff::{FromUniformBytes, WithSmallOrderMulGroup};
 
@@ -11,7 +11,9 @@ use crate::{
 };
 
 /// Given a plonk proof, this function parses it to extract the verifying trace.
-pub(crate) fn parse_trace<F, CS, T>(
+/// This function computes all Fiat-Shamir challenges, with the exception of
+/// `x`, which is computed in [verify_algebraic_constraints]
+pub fn parse_trace<F, CS, T>(
     vk: &VerifyingKey<F, CS>,
     // Unlike the prover, the verifier gets their instances in two arguments:
     // committed and normal (non-committed). Note that the total number of
@@ -169,11 +171,12 @@ where
     })
 }
 
-/// Given a [VerifierTrace], this function verifies the algebraic constraints
-/// with the claimed evaluations, but does not verify the PCS proof.
+/// Given a [VerifierTrace], this function computes the opening challenge, x,
+/// and proceeds to verify the algebraic constraints with the claimed
+/// evaluations. This function does not verify the PCS proof.
 ///
 /// The verifier will error if there are trailing bits in the transcript.
-pub(crate) fn verify_algebraic_constraints<F, CS: PolynomialCommitmentScheme<F>, T: Transcript>(
+pub fn verify_algebraic_constraints<F, CS: PolynomialCommitmentScheme<F>, T: Transcript>(
     vk: &VerifyingKey<F, CS>,
     trace: VerifierTrace<F, CS>,
     // Unlike the prover, the verifier gets their instances in two arguments:
@@ -190,6 +193,7 @@ where
         + Hashable<T::Hash>
         + Sampleable<T::Hash>
         + FromUniformBytes<64>
+        + Hash
         + Ord,
     CS::Commitment: Hashable<T::Hash>,
 {
@@ -217,9 +221,9 @@ where
     // Sample x challenge, which is used to ensure the circuit is
     // satisfied with high probability.
     let x: F = transcript.squeeze_challenge();
+    let xn = x.pow_vartime([vk.n()]);
 
     let instance_evals = {
-        let xn = x.pow([vk.n()]);
         let (min_rotation, max_rotation) =
             vk.cs
                 .instance_queries
@@ -303,9 +307,6 @@ where
     // This check ensures the circuit is satisfied so long as the polynomial
     // commitments open to the correct values.
     let vanishing = {
-        // x^n
-        let xn = x.pow([vk.n()]);
-
         let blinding_factors = vk.cs.blinding_factors();
         let l_evals = vk
             .domain
@@ -458,11 +459,12 @@ where
                 }),
         )
         .chain(permutations_common.queries(&vk.permutation, x))
-        .chain(vanishing.queries(x, vk.n()));
+        .chain(vanishing.queries(x, vk.n()))
+        .collect::<Vec<_>>();
 
     // We are now convinced the circuit is satisfied so long as the
     // polynomial commitments open to the correct values.
-    CS::multi_prepare(queries, transcript).map_err(|_| Error::Opening)
+    CS::multi_prepare(&queries, transcript).map_err(|_| Error::Opening)
 }
 
 /// Prepares a plonk proof into a PCS instance that can be finalized or
@@ -486,6 +488,7 @@ where
         + Hashable<T::Hash>
         + Sampleable<T::Hash>
         + FromUniformBytes<64>
+        + Hash
         + Ord,
     CS::Commitment: Hashable<T::Hash>,
 {